feat: moved conditionsInteraction file to components/ folder
